+++ test-data/p1-preloaded_lists/certification.sql
@@ -2,16 +2,12 @@ USE linkServicesGroup;
 
 INSERT INTO CheckClearance (checkClearanceName, checkClearanceDesc) VALUES
 ('Willing to obtain', 'This Support Worker is willing to obtain any checks or clearances necessary for a job role'),
-('Will Pass', 'This Support Worker will pass any checks or clearances necessary for a job role'),
 ('First Aid', 'This Support Worker has the skills and training required to perform First Aid'),
-('CPR', 'This Support Worker has the skills and training required to perform CPR'),
 ('Blue Card', 'This Support Worker is suitable for working with children and youth'),
 ('Criminal History Check', 'This Support Worker has undergone and passed a Criminal History check'),
-('Yellow Card', 'This Support Worker is suitable for working with vulnurable people of all types'),
 ('Police Check', 'This Support Worker has undergone and passed a Police check'),
 ('Insurance Check', 'This Support Worker has a required and up-to-date insurance policy'),
-('Full Comprehensive Car Insurance', 'This Support Worker has coverage to help pay to replace or repair their vehicle if it is stolen or damaged'),
-('Mental Health First Aid', 'This Support Worker has the skills necessary to assist people who are experiencing mental illness'),
+('Mental Health', 'This Support Worker has the skills necessary to assist people who are experiencing mental illness'),
 ('Hand Hygene and Infection Control', 'This Support Worker has knowledge regarding hand hygene practices and preventing the spread of infections'),
 ('Manual Handling', 'This Support Worker has the formal training and qualification necessary for manual handling and labour tasks'),
 ('Worker Orientation Module', 'This module is an interactive online course that explains the obligations of workers under the NDIS Code of Conduct.');