+++ changelog.md
@@ -1,4 +1,9 @@
 # Changelog
 
-**./data-dict.md**
-* Complete
+**./test-data/p1-preloaded_lists/certification.sql - CheckClearance**
+* Removed entries:
+	* 'Will Pass'
+	* 'Yellow Card'
+	* 'CPR'
+	* 'Full Comprehensive Car Insurance'
+* Renamed 'Mental Health First Aid' to ''Mental Health'